most of the admin dashboard part completed with secure data flow as i know

+++ app/Http/Controllers/RoomController.php
@@ -3,14 +3,16 @@
 namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Requests\RoomCreateRequest;
+use App\Models\Booking;
 use App\Models\Room;
 
 
 class RoomController extends Controller
 {
+
      public function index() {
         $rooms = Room::paginate(4);
-            return view('admin.dashboard.rooms.index',compact('rooms'));
+        return view('admin.dashboard.rooms.index',compact('rooms'));
     }
     public function create() {
         return view('admin.dashboard.rooms.create');
@@ -34,15 +36,18 @@ class RoomController extends Controller
         // dd('hi update');
         $updated_data = $request->validate([
             'room_type'   => 'required|string|',
-            'room_number' => 'required|string|max:3|unique:rooms,room_number',
-            'status'      => 'required|string|',
+            'room_number' => 'required|string|max:3|',
             'price'       => 'required|string',
         ]);
         if($room->update($updated_data)){
-            return redirect()->route('admin.show.rooms')->with('notification',['type'=>'success','message'=>'Room updated successfully']);
+            return redirect()->route('admin.show.rooms')->with('notification',['type'=>'success','message'=>'Room '. $room->room_number . ' updated successfully']);
         }
     }
     public function destroy(Room $room){
+        if($room->status == 'Booked'){
+            return redirect()->back()->with('notification',['type'=>'danger','message'=>'The room is booked. unable to delete.']);
+        }
+
         $deleted = $room->delete();
         if($deleted){
             return redirect()->route('admin.show.rooms')->with('notification',['type'=>'success','message'=>'Room deleted successfully']);
